[feat] inicio dos control
