+++ app.py
@@ -1,8 +1,8 @@
 from flask import Flask, jsonify, request
 from configuracao import app
-from controler.control_aluno import alunos_blueprint
+from aluno.control_aluno import aluno_blueprint
 
-app.register_blueprint(alunos_blueprint)
+app.register_blueprint(aluno_blueprint)
 
 dicionario = {
     "Alunos": [
@@ -145,31 +145,6 @@ def deleteTurma(idTurma):
     else:
         return jsonify({"erro": "Turma não encontrada"}),400
 
-#funcoes
-def calcula_idade(data):   
-    data = data.split("-")  
-    ano = int(data[0])      
-    mes = int(data[1])      
-    dia = int(data[2])      
-    data_nascimento = datetime.date(ano, mes, dia) 
-    hoje = datetime.date.today()
-    idade = hoje - data_nascimento
-    anos = idade.days // 365 
-    return anos
-
-def media(nota1, nota2):
-    soma = nota1 + nota2
-    return soma / 2
-
-def empty(texto):                       
-    texto = texto.capitalize()          
-    if texto in dicionario:             
-        if len(dicionario[texto]) > 0:  
-            return False                
-        else:                           
-            return True                 
-    else:
-        return False
 
 if __name__ == '__main__':
-    app.run(debug=True)
+    app.run(host=app.config["HOST"], port = app.config['PORT'],debug=app.config['DEBUG'] )
