+++ professor/model_professor.py
@@ -1,3 +1,4 @@
+#from funcoes import calcula_idade, empty
 import datetime
 
 dicionario = {
@@ -88,6 +89,10 @@ def delete_professor(idProfessor):
         return "Professor não encontrado"
 
 
+#funcoes
+def existe_professor():
+    return dicionario
+
 def calcula_idade(data):   
     data = data.split("-")  
     ano = int(data[0])      
@@ -98,3 +103,17 @@ def calcula_idade(data):
     idade = hoje - data_nascimento
     anos = idade.days // 365 
     return anos
+
+def media(nota1, nota2):
+    soma = nota1 + nota2
+    return soma / 2
+
+def empty(texto, dicionario):                       
+    texto = texto.capitalize()          
+    if texto in dicionario:             
+        if len(dicionario[texto]) > 0:  
+            return False                
+        else:                           
+            return True                 
+    else:
+        return False
